refactor(frontend): split automation inventory domain rules
